refactor: extract storage setters; move events to system
