+++ frame/system/system.go
@@ -7,7 +7,6 @@ import (
 	sc "github.com/LimeChain/goscale"
 	"github.com/LimeChain/gosemble/constants"
 	"github.com/LimeChain/gosemble/constants/system"
-	"github.com/LimeChain/gosemble/frame"
 	"github.com/LimeChain/gosemble/frame/timestamp"
 	"github.com/LimeChain/gosemble/primitives/hashing"
 	"github.com/LimeChain/gosemble/primitives/storage"
@@ -274,79 +273,8 @@ func EnsureInherentsAreFirst(block types.Block) int {
 	return -1
 }
 
-// The current block number being processed. Set by `execute_block`.
-func StorageGetBlockNumber() types.BlockNumber {
-	systemHash := hashing.Twox128(constants.KeySystem)
-	numberHash := hashing.Twox128(constants.KeyNumber)
-	return storage.GetDecode(append(systemHash, numberHash...), sc.DecodeU32)
-}
-
-func StorageGetAllExtrinsicsLen() sc.U32 {
-	systemHash := hashing.Twox128(constants.KeySystem)
-	allExtrinsicsLenHash := hashing.Twox128(constants.KeyAllExtrinsicsLen)
-	return storage.GetDecode(append(systemHash, allExtrinsicsLenHash...), sc.DecodeU32)
-}
-
-func StorageGetAccount(who types.PublicKey) types.AccountInfo {
-	systemHash := hashing.Twox128(constants.KeySystem)
-	accountHash := hashing.Twox128(constants.KeyAccount)
-
-	whoBytes := sc.FixedSequenceU8ToBytes(who)
-
-	key := append(systemHash, accountHash...)
-	key = append(key, hashing.Blake128(whoBytes)...)
-	key = append(key, whoBytes...)
-
-	return storage.GetDecode(key, types.DecodeAccountInfo)
-}
-
-func StorageSetAccount(who types.PublicKey, account types.AccountInfo) {
-	systemHash := hashing.Twox128(constants.KeySystem)
-	accountHash := hashing.Twox128(constants.KeyAccount)
-
-	whoBytes := sc.FixedSequenceU8ToBytes(who)
-
-	key := append(systemHash, accountHash...)
-	key = append(key, hashing.Blake128(whoBytes)...)
-	key = append(key, whoBytes...)
-
-	storage.Set(key, account.Bytes())
-}
-
-// Map of block numbers to block hashes.
-func StorageGetBlockHash(blockNumber sc.U32) types.Blake2bHash {
-	// Module prefix
-	systemHash := hashing.Twox128(constants.KeySystem)
-	// Storage prefix
-	blockHashHash := hashing.Twox128(constants.KeyBlockHash)
-	// Block number hash
-	blockNumHash := hashing.Twox64(blockNumber.Bytes())
-
-	key := append(systemHash, blockHashHash...)
-	key = append(key, blockNumHash...)
-	key = append(key, blockNumber.Bytes()...)
-
-	return storage.GetDecode(key, types.DecodeBlake2bHash)
-}
-
-// Map of block numbers to block hashes.
-func StorageExistsBlockHash(blockNumber sc.U32) sc.Bool {
-	// Module prefix
-	systemHash := hashing.Twox128(constants.KeySystem)
-	// Storage prefix
-	blockHashHash := hashing.Twox128(constants.KeyBlockHash)
-	// Block number hash
-	blockNumHash := hashing.Twox64(blockNumber.Bytes())
-
-	key := append(systemHash, blockHashHash...)
-	key = append(key, blockNumHash...)
-	key = append(key, blockNumber.Bytes()...)
-
-	return storage.Exists(key) == 1
-}
-
 func onCreatedAccount(who types.PublicKey) {
 	// hook on creating new account, currently not used in Substrate
 	//T::OnNewAccount::on_new_account(&who);
-	frame.DepositEvent(NewEventNewAccount(who), StorageGetBlockNumber())
+	DepositEvent(NewEventNewAccount(who))
 }
